+++ src/python/deepgnn/BUILD
@@ -10,7 +10,6 @@ py_library(
         "arg_types.py",
         "log_consts.py",
         "logging_utils.py",
-        "train_types.py",
     ],
     visibility = ["//visibility:public"],
 )
